requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
Faker
PyMySQL
streamlit
plotly-express